Clean up output
Don’t add a marker attribute to portals because it’ll be included in
the output.  Instead keep a dictionary of markers by guid.
Also, made the text entry boxes a little less than 100% so the labels
and buttons line up nicely.  At 100% a scroll bar gets added (hence the
little less).