Moved outer triangle selection into the app
Instead of making the user (well, me) deal with selecting the outer
triangle using the Ingress link planning tool (which is really
frustrating on Android), allow it to be defined by clicking on the
portal info boxes.  The problem with the Niantic tool is that when you
click on a portal to get the info box, the bit to click on is generally
off screen.  Scrolling to it generally results in the portal being
deselected and the info box being dismissed.  So anyway, this simpler
selection mechanism works a lot better.
Also, save the fields in the IndexedDB and add a select field drop down
to switch between the previously generated fields.  Field generation is
instantaneous, but it saves the time selecting the portals.
Maybe someday, I’ll need to add a way to get rid of old fields out of
the DB.  For now, this seems pretty much like v1.0.
Oh, and select_triangle.html still works and can be used as the entry
page, but it’s completely unnecessary.