Share link, caps and continuous spine
Added a “share” to share the field with others.
Added the ability to cap layers (7 is popular because a 7 layer field
with a continuous spine can be thrown from one of the anchors)
Added the ability to cap keys required (people might not want to farm
100 keys)
Added the ability to cap max keys per portal (people might not want to
farm 30 keys from a portal)
Added the ability to require a continuous spine.
TODO: Fix area calculation.  It’s pretending lat,lng are linear
cartesian coordinates.  So it favors tall (or is it wide?) fields.
TODO: Report area.  Sometimes the continuous spine can give more total
area because more fields are generated.  It’d be nice to show the sizes
(in km^2).
TO-PROBABLY-NOT-DO: Exhaustive search for the best sum of area by
trying all possibilities.
TO-MAYBE-DO: Search for the largest triangle containing no other points
and recurse outward to the bounding triangle.