Fixed a bug
Added a verbose flag to get enough log spam to discover that comparing
portals sometimes says != even though all the object attributes are
equal.  So compare guids instead.  Gotta be faster anyway.
Also, add a max of 999 to all the number inputs so they’re not so wide
on screen.