Less crowded
Swapped the meaning of E1 and neutral so that ignored portals are now
E1 and unused portals are neutral.
Hide all unused portals once the field has been found.
Fixed a bug in which the “select a portal” from “New…” field was still
displayed if “New…” was navigated away from via picking a field in the
drop down.