Portal removal
When importing from IITC, check for portals in the database in the same
rectangle (screen area in IITC) that aren’t part of the input and allow
those portals to be removed from the database.  If there’s never any
false positives, we can simplify it later to just remove them.